+++ examples/ssx-test-social/src/App.tsx
@@ -4,30 +4,22 @@ import { SSX } from '@spruceid/ssx';
 import './App.css';
 import getSSXConfig from './ssx.config';
 import AccountInfo from "./components/AccountInfo";
-
+import { ethers } from "ethers"
+import Login from "./components/Login";
 
 
 function App() {
 
   const [ssxProvider, setSSX] = useState<SSX | null>(null);
-  const API_KEY = process.env.API_KEY;
-  const PRIVATE_KEY = process.env.PRIVATE_KEY;
-  const CONTRACT_ADDRESS = process.env.CONTRACT_ADDRESS;
+  const API_KEY = process.env.REACT_APP_API_KEY;
+  const PRIVATE_KEY = process.env.REACT_APP_PRIVATE_KEY as any;
+  const CONTRACT_ADDRESS = process.env.REACT_APP_CONTRACT_ADDRESS as any;
   const contract = require("./artifacts/contracts/Posts.sol/Posts.json");
-  console.log(JSON.stringify(contract.abi));
+  const alchemyProvider = new ethers.providers.AlchemyProvider("goerli", API_KEY);
+  const signer = new ethers.Wallet(PRIVATE_KEY, alchemyProvider);
+  const PostsContract = new ethers.Contract(CONTRACT_ADDRESS, contract.abi, signer);
+  const [posts, setPosts] = useState([])
 
-  const ssxHandler = async () => {
-    const ssxConfig = await getSSXConfig();
-    const ssx = new SSX(ssxConfig);
-    await ssx.signIn();
-    setSSX(ssx);
-    (window as any).ssx = ssx;
-  };
-
-  const ssxLogoutHandler = async () => {
-    ssxProvider?.signOut();
-    setSSX(null);
-  };
 
   return (
     <div className="App">
@@ -39,22 +31,7 @@ function App() {
         />
         <span>SSX</span>
       </div>
-      <div className="App-content">
-        {
-          ssxProvider ?
-            <>
-              <button onClick={ssxLogoutHandler}>
-                SIGN OUT
-              </button>
-              <AccountInfo
-                address={ssxProvider?.address() || ''}
-              />
-            </> :
-            <button onClick={ssxHandler}>
-              SIGN-IN WITH ETHEREUM
-            </button>
-        }
-      </div>
+      <Login ssxProvider={ssxProvider} setSSX={setSSX}></Login>
     </div>
   );
 }
